+++ lib/dry/transaction/step_adapters/map.rb
@@ -3,10 +3,10 @@ module Dry
     class StepAdapters
       # @api private
       class Map
-        include Dry::Monads::Either::Mixin
+        include Dry::Monads::Result::Mixin
 
         def call(step, input, *args)
-          Right(step.call_operation(input, *args))
+          Success(step.call_operation(input, *args))
         end
       end