+++ lib/dry/transaction/step_adapters/raw.rb
@@ -1,17 +1,17 @@
-require "dry/monads/either"
+require "dry/monads/result"
 
 module Dry
   module Transaction
     class StepAdapters
       # @api private
       class Raw
-        include Dry::Monads::Either::Mixin
+        include Dry::Monads::Result::Mixin
 
         def call(step, input, *args)
           result = step.call_operation(input, *args)
 
-          unless result.is_a?(Dry::Monads::Either)
-            raise ArgumentError, "step +#{step.step_name}+ must return an Either object"
+          unless result.is_a?(Dry::Monads::Result)
+            raise ArgumentError, "step +#{step.step_name}+ must return a Result object"
           end
 
           result