+++ lib/dry/transaction/step_adapters/tee.rb
@@ -3,11 +3,11 @@ module Dry
     class StepAdapters
       # @api private
       class Tee
-        include Dry::Monads::Either::Mixin
+        include Dry::Monads::Result::Mixin
 
         def call(step, input, *args)
           step.call_operation(input, *args)
-          Right(input)
+          Success(input)
         end
       end