+++ spec/integration/publishing_step_events_spec.rb
@@ -4,7 +4,7 @@ RSpec.describe "publishing step events" do
       extend Dry::Container::Mixin
 
       register :process, -> input { {name: input["name"]} }
-      register :verify,  -> input { input[:name].to_s != "" ? Dry::Monads.Right(input) : Dry::Monads.Left("no name") }
+      register :verify,  -> input { input[:name].to_s != "" ? Dry::Monads.Success(input) : Dry::Monads.Failure("no name") }
       register :persist, -> input { Test::DB << input and true }
     end
   }
@@ -34,17 +34,17 @@ RSpec.describe "publishing step events" do
     specify "subscriber receives success events" do
       transaction.call("name" => "Jane")
 
-      expect(subscriber).to have_received(:process_success).with(name: "Jane")
-      expect(subscriber).to have_received(:verify_success).with(name: "Jane")
-      expect(subscriber).to have_received(:persist_success).with(name: "Jane")
+      expect(subscriber).to have_received(:step_succeeded).with(:process, "name" => "Jane")
+      expect(subscriber).to have_received(:step_succeeded).with(:verify, name: "Jane")
+      expect(subscriber).to have_received(:step_succeeded).with(:persist, name: "Jane")
     end
 
     specify "subsriber receives success events for passing steps, a failure event for the failing step, and no subsequent events" do
       transaction.call("name" => "")
 
-      expect(subscriber).to have_received(:process_success).with(name: "")
-      expect(subscriber).to have_received(:verify_failure).with({name: ""}, "no name")
-      expect(subscriber).not_to have_received(:persist_success)
+      expect(subscriber).to have_received(:step_succeeded).with(:process, "name" =>  "")
+      expect(subscriber).to have_received(:step_failed).with(:verify , {name: ""}, "no name")
+      expect(subscriber).not_to have_received(:step_succeeded).with(:persist)
     end
   end
 
@@ -56,15 +56,15 @@ RSpec.describe "publishing step events" do
     specify "subscriber receives success event for the specified step" do
       transaction.call("name" => "Jane")
 
-      expect(subscriber).to have_received(:verify_success).with(name: "Jane")
-      expect(subscriber).not_to have_received(:process_success)
-      expect(subscriber).not_to have_received(:persist_success)
+      expect(subscriber).to have_received(:step_succeeded).with(:verify, name: "Jane")
+      expect(subscriber).not_to have_received(:step_succeeded).with(:process)
+      expect(subscriber).not_to have_received(:step_succeeded).with(:persist)
     end
 
     specify "subscriber receives failure event for the specified step" do
       transaction.call("name" => "")
 
-      expect(subscriber).to have_received(:verify_failure).with({name: ""}, "no name")
+      expect(subscriber).to have_received(:step_failed).with(:verify, {name: ""}, "no name")
     end
   end
 
@@ -78,7 +78,7 @@ RSpec.describe "publishing step events" do
         extend Dry::Container::Mixin
 
         register :process, -> input { {name: input["name"]} }
-        register :verify,  -> input, name { input[:name].to_s == name ? Dry::Monads.Right(input) : Dry::Monads.Left("wrong name") }
+        register :verify,  -> input, name { input[:name].to_s == name ? Dry::Monads.Success(input) : Dry::Monads.Failure("wrong name") }
         register :persist, -> input { Test::DB << input and true }
       end
     }
@@ -86,15 +86,15 @@ RSpec.describe "publishing step events" do
     specify "subscriber receives success event for the specified step" do
       transaction.with_step_args(verify: ["Jane"]).call("name" => "Jane")
 
-      expect(subscriber).to have_received(:verify_success).with(name: "Jane")
-      expect(subscriber).not_to have_received(:process_success)
-      expect(subscriber).not_to have_received(:persist_success)
+      expect(subscriber).to have_received(:step_succeeded).with(:verify, {:name=>"Jane"}, "Jane")
+      expect(subscriber).not_to have_received(:step_succeeded).with(:process)
+      expect(subscriber).not_to have_received(:step_succeeded).with(:persist)
     end
 
     specify "subscriber receives failure event for the specified step" do
       transaction.with_step_args(verify: ["Jade"]).call("name" => "")
 
-      expect(subscriber).to have_received(:verify_failure).with({name: ""}, "Jade", "wrong name")
+      expect(subscriber).to have_received(:step_failed).with(:verify, {name: ""}, "Jade", "wrong name")
     end
   end
 end