+++ spec/integration/transaction_without_steps_spec.rb
@@ -41,11 +41,11 @@ RSpec.describe "Transactions steps without arguments" do
     end
 
     it "returns a success" do
-      expect(transaction.call()).to be_a Dry::Monads::Either::Right
+      expect(transaction.call()).to be_a Dry::Monads::Result::Success
     end
 
     it "wraps the result of the final operation" do
-      expect(transaction.call().value).to eq(name: "Jane", email: "jane@doe.com")
+      expect(transaction.call().value!).to eq(name: "Jane", email: "jane@doe.com")
     end
 
     it "supports matching on success" do